Assignment - 01 - Maciej Medyk
Hi, This is to get you started with government data. Use Python to
analyze the data here:
http://catalog.data.gov/dataset/data-gov-visitor-metrics . Click on the
Download button to download the .csv file. Aggregate the data by
organization. For the date period given in the file, create and pretty
print a table to show the top ten organizations (in terms of total views
of the data during this period). Plot a barchart. Submit a .ipynb file
at Github:https://github.com/Semantic-Web .I will soon start populating.
I need your UNs for Github ASAP, whether you plan to do this assignment
or not. I strongly urge you to give this a try - and ask for help at
LinkedIn. It is OK - one of us will help. The goal is to help each other
- your opportunity to contribute and help others will come.

Note that .ipynb is a 'literate programming' type of document - with
text, code, and results integrated, as shown in the class and as in the
.ipynb file attached to the Tutorial section's entry for Python. Use
only Anaconda and not any advanced tool such as Jupyter (so I can open
the file and execute the code).